models: split monolithic models.py into topical submodules
Move SQLAlchemy class definitions out of the 2.4k-line app_core/models.py
into eight topical sibling modules sharing helpers via _models_base.  The
public surface is preserved verbatim — every class, helper, and re-exported
name still resolves through ``from app_core.models import …`` exactly as
before, so no call site needs to change.
